pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dinobase"
version = "0.1.0"
description = "The agent-native database"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "duckdb>=1.0",
    "dlt[duckdb]>=1.0",
    "mcp>=1.0",
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
postgres = ["dlt[sql-database]"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
dinobase = "dinobase.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/dinobase"]

[tool.pytest.ini_options]
testpaths = ["tests"]
